Small UI Fixes
Fixed the image file drop box to be dynamic to the image placed. Fixed the "Sign In" button to match the style of the "Sign Up" Button. Updated the "Wardrobe" font size for each clothing article. Implemented a button to "Fabriq." logo to redirect the user to the home page.